(PC-23190)[API] chore: extract post product by ean test
